fix: move indexing-status lib to api/lib
